# extracted_specs/openapi_golden_congno.yaml
openapi: 3.0.3
info:
  title: Cong No Golden API
  version: "1.0.0"
servers:
  - url: http://server1:8080/api
...
paths:
...
  /reports/summary:
    get:
      tags: [Reports]
      summary: Summary report
      parameters:
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
        - in: query
          name: groupBy
          schema:
            type: string
            enum: [customer, owner, seller, period]
      responses:
        "200": { description: OK }
...